Sa8 32bit scales

+++ include/api/mli_helpers_api.h
@@ -132,7 +132,7 @@ mli_status mli_hlp_create_subtensor(const mli_tensor *in, const mli_sub_tensor_c
 
 uint32_t mli_hlp_tensor_scale_shift(const mli_tensor *in);
 
-int16_t mli_hlp_tensor_scale(const mli_tensor *in, const uint32_t scale_idx);
+int32_t mli_hlp_tensor_scale(const mli_tensor *in, const uint32_t scale_idx);
 
 int16_t mli_hlp_tensor_zero_offset(const mli_tensor *in, const uint32_t zero_idx);
